+++ core.php
@@ -650,7 +650,7 @@ class page_renderer
 			"{sitename}" => $logo_html,
 			"{version}" => $version,
 			"{favicon-url}" => $settings->favicon,
-			"{header-html}" => self::get_css_as_html(),
+			"{header-html}" => self::get_header_html(),
 
 			"{navigation-bar}" => self::render_navigation_bar($settings->nav_links, $settings->nav_links_extra, "top"),
 			"{navigation-bar-bottom}" => self::render_navigation_bar($settings->nav_links_bottom, [], "bottom"),
@@ -699,6 +699,25 @@ class page_renderer
 		return self::render($title, $content, self::$minimal_content_template);
 	}
 	
+	public static function get_header_html()
+	{
+		global $settings;
+		$result = self::get_css_as_html();
+		
+		if(!empty($settings->enable_math_rendering))
+			$result .= "<script type='text/x-mathjax-config'>
+  MathJax.Hub.Config({
+    tex2jax: {
+      inlineMath: [ ['$','$'], ['\\\\(','\\\\)'] ],
+      processEscapes: true,
+      skipTags: ['script','noscript','style','textarea','pre','code']
+    }
+  });
+</script>
+<script async src='https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-MML-AM_CHTML'></script>";
+		
+		return $result;
+	}
 	public static function get_css_as_html()
 	{
 		global $settings;